API Product support for updating swagger from dependent APIs

+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/APIGatewayManager.java
@@ -463,6 +463,31 @@ public class APIGatewayManager {
 
         return failedEnvironmentsMap;
     }
+
+    public Map<String, String> updateLocalEntry(APIProduct apiProduct, String tenantDomain) {
+        Map<String, String> failedEnvironmentsMap = new HashMap<>();
+
+        for (String environmentName : apiProduct.getEnvironments()) {
+
+            Environment environment = environments.get(environmentName);
+            try {
+                LocalEntryAdminClient localEntryAdminClient = new LocalEntryAdminClient(environment, tenantDomain);
+
+                String definition = apiProduct.getDefinition();
+                localEntryAdminClient.deleteEntry(apiProduct.getUuid());
+                localEntryAdminClient.addLocalEntry("<localEntry key=\"" + apiProduct.getUuid() + "\">" +
+                        definition.replaceAll("&(?!amp;)", "&amp;").
+                                replaceAll("<", "&lt;").replaceAll(">", "&gt;")
+                        + "</localEntry>");
+            } catch (AxisFault e) {
+                failedEnvironmentsMap.put(environmentName, e.getMessage());
+                log.error("Error occurred when publish to gateway " + environmentName, e);
+            }
+        }
+
+        return failedEnvironmentsMap;
+    }
+
 	/**
 	 * Removed an API from the configured Gateways
 	 * 
@@ -573,7 +598,13 @@ public class APIGatewayManager {
                     APIGatewayAdminClient client = new APIGatewayAdminClient(environment);
 
                     APIIdentifier id = new APIIdentifier(PRODUCT_PREFIX, apiProduct.getId().getName(), PRODUCT_VERSION);
-                    client.deleteApi(tenantDomain, id);
+                    // Endpoint may have been deleted already due to previously initiated API Product delete call that
+                    // only partially completed due an error that could have occurred at registry or DB level.
+                    if (client.getApi(tenantDomain, id) != null) {
+                        client.deleteApi(tenantDomain, id);
+                    } else {
+                        continue;
+                    }
 
                     for (API api : associatedAPIs) {
                         if (client.getApi(tenantDomain, api.getId()) == null) {
